+++ include/FTrack/Crit2_RZRatio.h
@@ -14,7 +14,7 @@ namespace FTrack{
 
    public:
       
-      Crit2_RZRatio ( double ratioMin, double ratioMax );
+      Crit2_RZRatio ( float ratioMin, float ratioMax );
       
       virtual bool areCompatible( Segment* parent , Segment* child )throw( BadSegmentLength );
 
@@ -22,8 +22,8 @@ namespace FTrack{
     
    private:
       
-      double _ratioMax;
-      double _ratioMin;
+      float _ratioMax;
+      float _ratioMin;
       
       
    };
